requirements.txt
anyio==4.11.0
certifi==2025.10.5
charset-normalizer==3.4.4
decorator==5.2.1
greenlet==3.1.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
ImageIO==2.37.2
imageio-ffmpeg==0.6.0
importlib_resources==6.5.2
moviepy==2.2.1
notion-client>=2.2.1
numpy==2.3.4
openai>=1.0.0
packaging==25.0
pillow==11.3.0
playwright==1.48.0
proglog==0.1.12
pydantic==1.10.13
pyee==12.0.0
python-dotenv==1.2.1
replicate==0.25.2
requests>=2.31.0
sniffio==1.3.1
tqdm==4.67.1
tweepy>=4.14.0
typeshed_client==2.8.2
typing_extensions==4.15.0
urllib3==2.5.0
